account form
the account form was created using the the useForm hook along with zod for validation which uses the insertAccountSchema from drizzle. added props that are used in the form. the form itself for now displays the input for the name and two buttons to submit and delete .

it is then passed into the new account sheet file