row action & conditional render
the row action was applied to the end of the row as a button. when user clicks the button a dropdown menu appears and gives user the option to edit. the useopen account hooks opens the sheet and the id is passed within the sheet. the edit account sheet renders the account form with prerendered data based on the conditional render.